Tasks 2
Consists of 2 forms and 2 functions:
form 1 : takes the name of the file, opens it, calls the Input_data function in the SimulationSystem class to fill the lists with data.
simulate function : fills the simulation tabl with data while simulating the system.
form Simulation table: fills the data grid view with the data from the simulation table list.